+++ packages/kysely-driver/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @powersync/kysely-driver
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [395ea24]
+- Updated dependencies [9d1dc6f]
+  - @powersync/common@1.8.0
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [3c421ea]
+  - @powersync/common@1.7.1
+
 ## 0.2.0
 
 ### Minor Changes